backend/requirements.txt
asgiref>=3.8.1
Django>=4.2,<5.3
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.1
gunicorn==22.0.0
whitenoise==6.7.0
dj-database-url==2.1.0
django-storages==1.14.4
boto3==1.34.14drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.11.1
